try pass test

+++ controllers/cards.js
@@ -1,5 +1,6 @@
 const Card = require('../models/card');
 const NotFoundError = require('../errors/NotFoundError');
+const ForbidenErrors = require('../errors/ForbidenErrors');
 
 const getCards = (_, res, next) => {
   Card.find({})
@@ -15,10 +16,24 @@ const createCard = (req, res, next) => {
     .catch(next);
 };
 
+// const deleteCard = (req, res, next) => {
+//   Card.findByIdAndRemove(req.params.cardId)
+//     .orFail(new NotFoundError())
+//     .then((card) => res.send(card))
+//     .catch(next);
+// };
+
 const deleteCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .orFail(new NotFoundError())
-    .then((card) => res.send(card))
+    .populate('owner')
+    .then((card) => {
+      if (card.owner._id.toString() === req.user._id) {
+        return Card.findByIdAndRemove(req.params.cardId)
+          .then((deletedCard) => res.send(deletedCard));
+      }
+      return new ForbidenErrors();
+    })
     .catch(next);
 };
